Added Cluster Grammar Repo
This repo takes the mutfeats_ovarian... and mutations as main input, then uses wnthumans, kmeans & DDR genes to tag the genes with features and format it in a way that is cluster-grammar friendly in a jupyter notebook python file "clustergrammar-ncidata.ipynb". The resulting nci-featdat.txt can be loaded into the cluster-grammar website using jaccard to look at the relation between mutfeats and CIMP+ identifiers located in mutations.csv. The path for the files is hardcoded, and will have to be re-written upon download for whatever path you take to those files. I might want to change the paths in the future to have a designated input and output file so as not to confuse myself.